lib: Rename lib/api_test_helpers.py to openapi/python_examples.py
This will make the contained code easier to find.
