+++ tools/test-backend
@@ -109,7 +109,7 @@ not_yet_fully_covered = {path for target in [
     'zerver/lib/generate_test_data.py',
     'zerver/lib/test_fixtures.py',
     'zerver/lib/test_runner.py',
-    'zerver/lib/api_test_helpers.py',
+    'zerver/openapi/python_examples.py',
     # Tornado should ideally have full coverage, but we're not there.
     'zerver/tornado/autoreload.py',
     'zerver/tornado/descriptors.py',
